Fix SIP arrangement when pkgs share the same name
This commit fixes SIP arrangement when two or more SIPs are using the same
name. It ensures that the SIP object that we store in the database is using the
new path that we have assigned after the name clash was identified, i.e.
"../Foobar_1" as opposed to "../Foobar". Not doing this was causing the
MCPServer to create a whole new SIP, cuasing the `MultipleObjectsReturned`
error seen in our logs.

It also introduces a change that shows a warning to the user when an
empty SIP is being arranged since that is not a workflow that we
support.