Add new format for info
This commit introduces the desired format and saves
only the information which are wanted for sip, file and
event objects.
It must be optimised causing a lot of unperformant inner loops,
which will be a big performance problem running bigger datasets.
Collection name returns always null which might be related to the fact
that the aip_filename is generated and stored in the db after the
function is calling it from the db.